Создал вспомогательный дата-класс MovieVO.
