+++ app/src/main/java/ru/androidschool/intensiv/data/dto/BelongsToCollection.kt
@@ -1,4 +1,4 @@
-package ru.androidschool.intensiv.data
+package ru.androidschool.intensiv.data.dto
 
 import com.google.gson.annotations.SerializedName
 import ru.androidschool.intensiv.BuildConfig
