+++ app/src/main/java/ru/androidschool/intensiv/data/dto/Crew.kt
@@ -1,22 +1,21 @@
-package ru.androidschool.intensiv.data
+package ru.androidschool.intensiv.data.dto
 
 import com.google.gson.annotations.SerializedName
 import ru.androidschool.intensiv.BuildConfig
 
-// класс, обозначающий актера
-
-data class Actor(
+data class Crew(
     @SerializedName("adult") val isAdult: Boolean,
-    @SerializedName("gender") val gender: Int? = null,
+    @SerializedName("gender") val gender: Int? = null, // согласно документации integer or null
     @SerializedName("id") val id: Int,
     @SerializedName("known_for_department") val knownForDepartment: String,
     @SerializedName("name") val name: String,
     @SerializedName("original_name") val originalName: String,
     @SerializedName("popularity") val popularity: Double,
-    @SerializedName("character") val character: String,
     @SerializedName("credit_id") val creditId: String,
-    @SerializedName("order") val order: Int,
-    @SerializedName("profile_path") val profilePath: String? = null // согласно документации string or null
+    @SerializedName("department") val department: String,
+    @SerializedName("job") val job: String
 ) {
-    fun getProfilePathWithImageUrl(): String = "${BuildConfig.IMAGE_URL}$profilePath"
+    @SerializedName("profile_path")
+    val profilePath: String? = null // согласно документации string or null
+        get() = "${BuildConfig.IMAGE_URL}$field"
 }
