+++ app/src/main/java/ru/androidschool/intensiv/data/dto/LastEpisodeToAir.kt
@@ -1,4 +1,4 @@
-package ru.androidschool.intensiv.data
+package ru.androidschool.intensiv.data.dto
 
 import com.google.gson.annotations.SerializedName
