+++ app/src/main/java/ru/androidschool/intensiv/data/dto/Movie.kt
@@ -1,22 +1,22 @@
-package ru.androidschool.intensiv.data
+package ru.androidschool.intensiv.data.dto
 
 import com.google.gson.annotations.SerializedName
 import ru.androidschool.intensiv.BuildConfig
 
-class TvShow(
-    @SerializedName("first_air_date") val firstAirDate: String,
+data class Movie(
+    @SerializedName("adult") val isAdult: Boolean,
     @SerializedName("genre_ids") val genreIds: List<Int>,
     @SerializedName("id") val id: Int,
-    @SerializedName("name") val name: String,
-    @SerializedName("origin_country") val originCountry: List<String>,
     @SerializedName("original_language") val originalLanguage: String,
-    @SerializedName("original_name") val originalName: String,
+    @SerializedName("original_title") val originalTitle: String,
     @SerializedName("overview") val overview: String,
     @SerializedName("popularity") val popularity: Double,
+    @SerializedName("release_date") val releaseDate: String,
+    @SerializedName("title") val title: String,
+    @SerializedName("video") val video: Boolean,
     @SerializedName("vote_average") val voteAverage: Double,
     @SerializedName("vote_count") val voteCount: Int,
     @SerializedName("poster_path") val posterPath: String? = null // согласно документации string or null
-
 ) {
     val rating: Float
         get() = voteAverage.div(2).toFloat()
