+++ app/src/main/java/ru/androidschool/intensiv/data/dto/MovieCreditsResponse.kt
@@ -1,6 +1,8 @@
-package ru.androidschool.intensiv.data
+package ru.androidschool.intensiv.data.dto
 
 import com.google.gson.annotations.SerializedName
+import ru.androidschool.intensiv.data.dto.Actor
+import ru.androidschool.intensiv.data.dto.Crew
 
 data class MovieCreditsResponse(
     @SerializedName("id") val id: Int,
