+++ app/src/main/java/ru/androidschool/intensiv/data/dto/MovieDetails.kt
@@ -1,36 +1,30 @@
-package ru.androidschool.intensiv.data
+package ru.androidschool.intensiv.data.dto
 
 import com.google.gson.annotations.SerializedName
 import ru.androidschool.intensiv.BuildConfig
 
-data class TvShowDetails(
-    @SerializedName("created_by") val createdBy: List<CreatedBy>,
-    @SerializedName("episode_run_time") val episodeRunTime: List<Int>,
-    @SerializedName("first_air_date") val firstAirDate: String,
+data class MovieDetails(
+    @SerializedName("adult") val isAdult: Boolean,
+    @SerializedName("belongs_to_collection") val belongsToCollection: BelongsToCollection,
+    @SerializedName("budget") val budget: Int,
     @SerializedName("genres") val genres: List<Genre>,
     @SerializedName("homepage") val homepage: String,
     @SerializedName("id") val id: Int,
-    @SerializedName("in_production") val inProduction: Boolean,
-    @SerializedName("languages") val languages: List<String>,
-    @SerializedName("last_air_date") val lastAirDate: String,
-    @SerializedName("last_episode_to_air") val lastEpisodeToAir: LastEpisodeToAir,
-    @SerializedName("name") val name: String,
-    @SerializedName("next_episode_to_air") val nextEpisodeToAir: String,
-    @SerializedName("networks") val networks: List<Network>,
-    @SerializedName("number_of_episodes") val numberOfEpisodes: Int,
-    @SerializedName("number_of_seasons") val numberOfSeasons: Int,
-    @SerializedName("origin_country") val originCountry: List<String>,
+    @SerializedName("imdb_id") val imdbId: String,
     @SerializedName("original_language") val originalLanguage: String,
-    @SerializedName("original_name") val originalName: String,
+    @SerializedName("original_title") val originalTitle: String,
     @SerializedName("overview") val overview: String,
     @SerializedName("popularity") val popularity: Double,
     @SerializedName("production_companies") val productionCompanies: List<ProductionCompany>,
     @SerializedName("production_countries") val productionCountries: List<ProductionCountry>,
-    @SerializedName("seasons") val seasons: List<Season>,
+    @SerializedName("release_date") val releaseDate: String,
+    @SerializedName("revenue") val revenue: Int,
+    @SerializedName("runtime") val runtime: Int,
     @SerializedName("spoken_languages") val spokenLanguages: List<SpokenLanguage>,
     @SerializedName("status") val status: String,
     @SerializedName("tagline") val tagline: String,
-    @SerializedName("type") val type: String,
+    @SerializedName("title") val title: String,
+    @SerializedName("video") val video: Boolean,
     @SerializedName("vote_average") val voteAverage: Double,
     @SerializedName("vote_count") val voteCount: Int,
     @SerializedName("poster_path") val posterPath: String? = null // согласно документации string or null
