+++ app/src/main/java/ru/androidschool/intensiv/data/dto/TvShowResponse.kt
@@ -1,19 +1,17 @@
-package ru.androidschool.intensiv.data
+package ru.androidschool.intensiv.data.dto
 
 import com.google.gson.annotations.SerializedName
 
-data class SearchResponse(
-
+data class TvShowResponse(
     @SerializedName("page")
     val page: Int,
 
     @SerializedName("results")
-    val results: List<Movie>,
+    val results: List<TvShow>,
 
     @SerializedName("total_pages")
     val totalPages: Int,
 
     @SerializedName("total_results")
     val totalResults: Int
-
 )
