+++ app/src/main/java/ru/androidschool/intensiv/ui/feed/MovieItem.kt
@@ -4,7 +4,7 @@ import com.xwray.groupie.kotlinandroidextensions.GroupieViewHolder
 import com.xwray.groupie.kotlinandroidextensions.Item
 import kotlinx.android.synthetic.main.item_with_text.*
 import ru.androidschool.intensiv.R
-import ru.androidschool.intensiv.data.Movie
+import ru.androidschool.intensiv.data.dto.Movie
 import ru.androidschool.intensiv.ui.loadImage
 
 class MovieItem(
