+++ app/src/main/java/ru/androidschool/intensiv/ui/tvshows/TvShowItem.kt
@@ -4,7 +4,7 @@ import com.xwray.groupie.kotlinandroidextensions.GroupieViewHolder
 import com.xwray.groupie.kotlinandroidextensions.Item
 import kotlinx.android.synthetic.main.item_tv_show.*
 import ru.androidschool.intensiv.R
-import ru.androidschool.intensiv.data.TvShow
+import ru.androidschool.intensiv.data.dto.TvShow
 import ru.androidschool.intensiv.ui.loadImage
 
 class TvShowItem(
