+++ app/src/main/java/ru/androidschool/intensiv/ui/watchlist/MoviePreviewItem.kt
@@ -4,7 +4,6 @@ import com.xwray.groupie.kotlinandroidextensions.GroupieViewHolder
 import com.xwray.groupie.kotlinandroidextensions.Item
 import kotlinx.android.synthetic.main.item_with_text.*
 import ru.androidschool.intensiv.R
-import ru.androidschool.intensiv.data.Movie
 import ru.androidschool.intensiv.data.dbo.MovieDBO
 import ru.androidschool.intensiv.ui.loadImage
